refactor: pascal-case to kebab-case
